backend/requirements.txt
# Flask and web server
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0

# YouTube audio extraction
pytube==15.0.0

# Speech recognition
openai-whisper==20231117
ffmpeg-python==0.2.0

# Transformers for summarization
transformers==4.35.2
torch==2.1.0
sentencepiece==0.1.99

# Text processing and summarization
sumy==0.11.0
nltk==3.8.1

# Utilities
numpy==1.26.0
tqdm==4.66.1